Change implementation of AutoMapper and tidy up
- Add AutoMapper DI package
- Add UHPropertyProfile for AutoMapper

- Remove EntityFactory.cs
- Remove EntityMapper.cs
- Remove PropertyHelper.cs

- Split currently disabled logic for multiple properties out to separate file
- Rename PropertyInformationApiController to GetPropertyInformationController
- Rename GetProperty to GetPropertyUseCase
- Rename IGetProperty to IGetPropertyUseCase
