--- PropertyInformationApi/V1/Factories/EntityFactory.cs
@@ -1,20 +1,0 @@
-using PropertyInformationApi.V1.Domain;
-using PropertyInformationApi.V1.Infrastructure;
-
-namespace PropertyInformationApi.V1.Factories
-{
-    public static class EntityFactory
-    {
-        public static Entity ToDomain(this DatabaseEntity databaseEntity)
-        {
-            //TODO: Map the rest of the fields in the domain object.
-            // More information on this can be found here https://github.com/LBHackney-IT/lbh-base-api/wiki/Factory-object-mappings
-
-            return new Entity
-            {
-                Id = databaseEntity.Id,
-                CreatedAt = databaseEntity.CreatedAt,
-            };
-        }
-    }
-}
