+++ PropertyInformationApi/V1/UseCase/Interfaces/IGetPropertyUseCase.cs
@@ -2,7 +2,7 @@ using PropertyInformationApi.V1.Domain;
 
 namespace PropertyInformationApi.V1.UseCase
 {
-    public interface IGetProperty
+    public interface IGetPropertyUseCase
     {
         HousingProperty Execute(string propertyReference);
     }
